index

+++ 0x00-MySQL_Advanced/10-div.sql
@@ -1,6 +1,6 @@
 -- function
 DELIMITER $$
-CREATE FUNCTION SafeDiv(a int, b int) RETURNS DECIMAL(10, 2)
+CREATE FUNCTION SafeDiv(a int, b int) RETURNS DECIMAL
 BEGIN
 	RETURN IF (b != 0, a / b, 0);
 END $$
